requirements.txt
# Core dependencies
evds  # Resmi TCMB EVDS API paketi (5 Nisan 2024 güncellemesi ile uyumlu)
pandas
python-dotenv
toml
pykap  # KAP (Kamuyu Aydınlatma Platformu) API paketi

# Data processing
numpy

# Utilities
requests

# Development (optional)
jupyter
matplotlib
